+++ state_beacon/lib/src/base_beacon.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:math';
+import 'package:flutter/foundation.dart';
 import 'package:state_beacon/src/interfaces.dart';
 
 import 'async_value.dart';
@@ -20,11 +21,12 @@ part 'beacons/stream.dart';
 part 'beacons/list.dart';
 part 'beacons/derived.dart';
 part 'beacons/derived_future.dart';
+part 'beacons/value_notifier.dart';
 
 typedef VoidCallback = void Function();
 typedef Listerners = Set<EffectClosure>;
 
-abstract class BaseBeacon<T> {
+abstract class BaseBeacon<T> implements ValueListenable<T> {
   BaseBeacon([T? initialValue]) {
     if (initialValue != null || isNullable) {
       _initialValue = initialValue as T;
@@ -43,6 +45,7 @@ abstract class BaseBeacon<T> {
 
   T? get previousValue => _previousValue;
 
+  @override
   T get value {
     if (_isEmpty) {
       throw UninitializeLazyReadException();
@@ -122,5 +125,18 @@ abstract class BaseBeacon<T> {
   void dispose() {
     listeners.clear();
     reset();
+  }
+
+  @override
+  void addListener(VoidCallback listener) {
+    final effectClosure = EffectClosure(listener, customID: listener.hashCode);
+
+    listeners.add(effectClosure);
+  }
+
+  @override
+  void removeListener(VoidCallback listener) {
+    final effectClosure = EffectClosure(listener, customID: listener.hashCode);
+    listeners.remove(effectClosure);
   }
 }