+++ state_beacon/test/beacon_test.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
+import 'package:flutter_test/flutter_test.dart';
 import 'package:state_beacon/src/base_beacon.dart';
 import 'package:state_beacon/state_beacon.dart';
-import 'package:test/test.dart';
 
 void main() {
   group('Beacon Tests', () {